% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_similarity.R, R/function.R
\name{average_similarity}
\alias{average_similarity}
\title{计算每个神经元类型的平均相似性}
\usage{
average_similarity(neuron_data)

average_similarity(neuron_data)
}
\arguments{
\item{neuron_data}{数据框，包含神经元的类型、索引和相似性。}
}
\value{
返回一个数据框，包含每个神经元类型及其平均相似性。

返回一个数据框，包含每个神经元类型及其平均相似性。
}
\description{
该函数根据每个神经元的相似性计算每个类型（ON、OFF、其他）的平均相似性。

该函数根据每个神经元的相似性计算每个类型（ON、OFF、其他）的平均相似性。
}
