% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_neurons.R
\name{classify_neurons}
\alias{classify_neurons}
\title{生成神经元类型索引与相似性数据}
\usage{
classify_neurons(trace_mat, result_table)
}
\arguments{
\item{trace_mat}{神经元的钙成像数据矩阵}

\item{result_table}{行为数据的表格}
}
\value{
返回包含神经元类型、索引和相似性的数据框
}
\description{
根据神经元的钙成像数据和行为数据，生成神经元类型索引及其相似性。
}
