% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R, R/identify_neurons.R
\name{identify_neurons}
\alias{identify_neurons}
\title{神经元类型识别}
\usage{
identify_neurons(Cn, B)

identify_neurons(Cn, B)
}
\arguments{
\item{Cn}{一个神经元的钙成像数据向量}

\item{B}{一个与行为相关的数据向量}
}
\value{
返回一个字符型变量，表示神经元的类型（"ON神经元"、"OFF神经元" 或 "其他神经元"）。

返回一个神经元类型的字符串（"ON神经元"、"OFF神经元" 或 "其他神经元"）
}
\description{
该函数根据神经元活动与随机重排的置换测试结果，识别神经元类型（ON、OFF 或 其他）。

根据实际相似性与随机相似性的比较，标识神经元类型为 "ON神经元"、"OFF神经元" 或 "其他神经元"。
}
