% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{neuron_classification}
\alias{neuron_classification}
\title{神经元分类与相似性计算}
\usage{
neuron_classification(trace_mat, result_table)
}
\arguments{
\item{trace_mat}{数值矩阵，每行表示一个神经元的活动。}

\item{result_table}{数据框，包含其他与神经元相关的表格数据（例如行为数据）。}
}
\value{
返回一个包含神经元类型、索引和相似性的完整数据框。
}
\description{
该函数通过对每个神经元的活动数据进行分析，自动分类神经元类型，并计算它们的相似性。分类结果及相似性被存储在数据框中。
}
