% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R, R/permutation_test.R
\name{permutation_test}
\alias{permutation_test}
\title{随机重排置换测试}
\usage{
permutation_test(Cn, B, iterations = 10000)

permutation_test(Cn, B, iterations = 10000)
}
\arguments{
\item{Cn}{一个神经元的钙成像数据向量}

\item{B}{一个与行为相关的数据向量}

\item{iterations}{置换测试的迭代次数，默认为 10000}
}
\value{
返回一个数值向量，表示每次置换后的相似性。

返回一个包含随机相似性值的向量
}
\description{
该函数执行随机重排的置换测试，评估两个神经元活动的相似性是否显著。

进行随机重排的置换测试，用于计算相似性分布。
}
