% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_neuron_rois.R
\name{plot_neuron_rois}
\alias{plot_neuron_rois}
\title{绘制分类神经元的 ROI 位置}
\usage{
plot_neuron_rois(classification_file, roi_folder_path, fill = FALSE)
}
\arguments{
\item{classification_file}{CSV 文件路径，包含神经元索引和分类标签}

\item{roi_folder_path}{ROI 文件所在的文件夹路径}

\item{fill}{是否填充 ROI，默认为 FALSE（空心），如果为 TRUE，则绘制实心}
}
\value{
绘制神经元 ROI 图
}
\description{
绘制分类神经元的 ROI 位置
}
