% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{sort_neurons_by_similarity}
\alias{sort_neurons_by_similarity}
\title{按照相似性排序神经元}
\usage{
sort_neurons_by_similarity(neuron_data)
}
\arguments{
\item{neuron_data}{数据框，包含神经元的类型、索引和相似性。}
}
\value{
返回一个包含按相似性排序后的神经元数据框。
}
\description{
该函数根据神经元的相似性对不同类型的神经元进行排序。
}
